feat: Revamp GitHub Actions Workflow (#1)
* feat: Revamp GitHub Actions Workflow

This commit includes an overhaul of the GitHub Actions workflow to include test, lint, and Docker job execution. Added Go version, GolangCI Lint version environmental variables, changed Docker Image name, and modified the Docker build and push operation. The workflow now triggers on push (tagged as v*.*.*) and pull request on the main branch. The docker job now depends on successful completion of test and lint jobs. It builds Docker on two platforms (linux/amd64, linux/arm64) and pushes if event is not a pull request.

* wip

* wip

* wip

* wip

* wip

* wip

* wip